packages/lambdas/requirements-dev.txt
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
black
boto3-stubs[s3]
licenseheaders
moto
mypy
projen==0.58.33
pytest==6.2.1
types-python-dateutil
